requirements.txt
discord
art
colorama
python-dotenv
PyNaCl
